Remove unused code and submodules.
